Getting back into the project: Movement updates, coding standards, crosshair
Coming back to the game after a bit. There seems to be an update related to IWYU. Updated QuarantineSimCharacter to follow IWYU standards.

Completely revamped movement code. Removed springarm for camera. First perso view functions as expected now.

Crouch has no animation, but now crouches very low for debugging.

Moved some assets that shouldn't be visible in the beginning of game out of the level.

Added a crosshair -- considering using Slate to handle HUD but don't know how to use it/what it is yet.

Would like to have the crosshair image change so that when interactables are detected, the crosshair changes.